Refactored GlobalFiltering into folder with separate base, types, and utils
