+++ packages/table-core/src/core/table.ts
@@ -31,7 +31,7 @@ import { ColumnPinning } from '../features/column-pinning/ColumnPinning'
 import { ColumnSizing } from '../features/column-sizing/ColumnSizing'
 import { ColumnVisibility } from '../features/column-visibility/ColumnVisibility'
 import { GlobalFaceting } from '../features/global-faceting/GlobalFaceting'
-import { GlobalFiltering } from '../features/GlobalFiltering'
+import { GlobalFiltering } from '../features/global-filtering/GlobalFiltering'
 import { RowExpanding } from '../features/RowExpanding'
 import { RowPagination } from '../features/RowPagination'
 import { RowPinning } from '../features/row-pinning/RowPinning'
